requirements.txt
autopep8==1.4.3
bleach==3.0.2
certifi==2018.11.29
chardet==3.0.4
DateTime==4.3
defusedxml==0.5.0
Django==2.1.5
django-allauth==0.38.0
django-rest-auth==0.9.3
djangorestframework==3.9.0
docutils==0.14
future==0.17.1
idna==2.7
jedi==0.13.2
mccabe==0.6.1
oauthlib==3.0.0
parso==0.3.1
pkginfo==1.4.2
pluggy==0.8.1
pycodestyle==2.4.0
pydocstyle==3.0.0
pyflakes==2.0.0
Pygments==2.2.0
python-dateutil==2.7.5
python-jsonrpc-server==0.1.0
python-language-server==0.21.5
python3-openid==3.1.0
pytz==2018.7
readme-renderer==24.0
requests==2.20.1
requests-oauthlib==1.2.0
requests-toolbelt==0.8.0
rope==0.11.0
six==1.12.0
snowballstemmer==1.2.1
tqdm==4.28.1
twine==1.12.1
urllib3==1.24.1
webencodings==0.5.1
yapf==0.25.0
zope.interface==4.6.0
